refactor: 将emotinPlayer和actionPlayer独立
